[#28 refactoring] 패키지 변경
