+++ feature/save/src/main/java/com/dorabangs/feature/save/screen/DoraLinkSaveRoute.kt
@@ -1,14 +1,17 @@
-package com.dorabangs.feature.save
+package com.dorabangs.feature.save.screen
 
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
+import androidx.hilt.navigation.compose.hiltViewModel
+import com.dorabangs.feature.save.DoraSaveViewModel
 
 @Composable
 fun DoraLinkSaveRoute(
     onClickBackIcon: () -> Unit,
     onClickSaveButton: () -> Unit,
     modifier: Modifier = Modifier,
+    viewModel: DoraSaveViewModel = hiltViewModel(),
 ) {
     DoraLinkSaveScreen(
         modifier = modifier,
